Spostati ulteriori metodi.
