+++ BusinessLayer/BL_AnswerManagement.cs
@@ -1,10 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Data.Common;
 using System.Text;
+using SchoolGrades.DbClasses;
 
 namespace SchoolGrades
 {
     internal partial class BusinessLayer
     {
+        
     }
 }
