+++ DbClasses/DbAndBusiness.cs
@@ -3122,148 +3122,8 @@ namespace SchoolGrades.DbClasses
             return a;
         }
 
-        internal List<Student> GetAllStudentsThatAnsweredToATest(Test Test, Class Class)
-        {
-            List<Student> list = new List<Student>();
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                string query = "SELECT DISTINCT StudentsAnswers.IdStudent" +
-                    " FROM StudentsAnswers" +
-                    " JOIN Classes_Students ON StudentsAnswers.IdStudent=Classes_Students.IdStudent" +
-                    " JOIN Students ON Classes_Students.IdStudent=Students.IdStudent" +
-                    " WHERE StudentsAnswers.IdTest=" + Test.IdTest + "" +
-                    " AND Classes_Students.IdClass=" + Class.IdClass + "" +
-                    " ORDER BY Students.LastName, Students.FirstName, Students.IdStudent " +
-                    ";";
-                cmd.CommandText = query;
-                DbDataReader dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    int? idStudent = SafeDb.SafeInt(dRead["idStudent"]);
-                    Student s = dl.GetStudent(idStudent);
-                    list.Add(s);
-                }
-            }
-            return list;
-        }
-       
-        internal List<Answer> GetAllCorrectAnswersToThisQuestionOfThisTest(int? IdQuestion, int? IdTest)
-        {
-            List<Answer> list = new List<Answer>();
-            using (DbConnection conn = dl.Connect())
-            {
-                DbDataReader dRead;
-                DbCommand cmd = conn.CreateCommand();
-                string query = "SELECT Answers.*" +
-                    " FROM Answers" +
-                    " JOIN Questions ON Questions.IdQuestion=Answers.IdQuestion" +
-                    " JOIN Tests_Questions ON Questions.IdQuestion=Tests_Questions.IdQuestion" +
-                    " WHERE Questions.IdQuestion=" + IdQuestion + "" +
-                    " AND Tests_Questions.IdTest=" + IdTest + "" +
-                    " ORDER BY idAnswer" +
-                    ";";
-                cmd.CommandText = query;
-                dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    Answer a = GetAnswerFromRow(dRead);
-                    list.Add(a);
-                }
-            }
-            return list;
-        }
         
-        internal Answer GetAnswerFromRow(DbDataReader Row)
-        {
-            Answer a = new Answer();
-            a.IdAnswer = SafeDb.SafeInt(Row["IdAnswer"]);
-            a.IdQuestion = SafeDb.SafeInt(Row["IdQuestion"]);
-            a.ShowingOrder = SafeDb.SafeInt(Row["ShowingOrder"]);
-            a.Text = SafeDb.SafeString(Row["Text"]);
-            a.ErrorCost = SafeDb.SafeInt(Row["ErrorCost"]);
-            a.IsCorrect = SafeDb.SafeBool(Row["IsCorrect"]);
-            a.IsOpenAnswer = SafeDb.SafeBool(Row["IsOpenAnswer"]);
-            a.IsMutex = SafeDb.SafeBool(Row["IsMutex"]);
-
-            return a;
-        }
-
-        internal int? SaveAnnotation(StudentAnnotation Annotation, Student s)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                string query = "";
-                // find if an answer has already been given
-                if (Annotation.IdAnnotation != null && Annotation.IdAnnotation != 0)
-                {   // update answer
-                    query = "UPDATE StudentsAnnotations" +
-                    " SET" +
-                    " idStudent=" + SqlVal.SqlInt(s.IdStudent) + "," +
-                    " idSchoolYear='" + SqlVal.SqlString(Annotation.IdSchoolYear) + "'," +
-                    " instantTaken=" + SqlVal.SqlDate(Annotation.InstantTaken) + "," +
-                    " instantClosed=" + SqlVal.SqlDate(Annotation.InstantClosed) + "," +
-                    " isActive=" + SqlVal.SqlBool(Annotation.IsActive) + "," +
-                    " annotation='" + SqlVal.SqlString(Annotation.Annotation) + "'" +
-                    " WHERE idStudent=" +  SqlVal.SqlInt(s.IdStudent) +
-                    ";";
-                }
-                else
-                {
-                    Annotation.InstantTaken = DateTime.Now;
-                    Annotation.IsActive = true; 
-                    // create answer on database
-                    int? nextId = dl.NextKey("StudentsAnnotations", "IdAnnotation");
-                    Annotation.IdAnnotation = nextId;
-
-                    query = "INSERT INTO StudentsAnnotations " +
-                    "(idAnnotation, idStudent, annotation,instantTaken," +
-                    "instantClosed,isActive";
-                    if (Annotation.IdSchoolYear != null && Annotation.IdSchoolYear != "")
-                        query += ",idSchoolYear";
-                    query += ")";
-                    query += " Values(";
-                    query += "" + SqlVal.SqlInt(Annotation.IdAnnotation) + ",";
-                    query += "" + SqlVal.SqlInt(s.IdStudent) + ",";
-                    query += "'" + SqlVal.SqlString(Annotation.Annotation) + "'";
-                    query += "," + SqlVal.SqlDate(Annotation.InstantTaken);
-                    query += "," + SqlVal.SqlDate(Annotation.InstantClosed);
-                    query += "," + SqlVal.SqlBool(Annotation.IsActive);
-                    if (Annotation.IdSchoolYear != null && Annotation.IdSchoolYear != "")
-                        query += ",'" + SqlVal.SqlString(Annotation.IdSchoolYear) + "'";
-                    query += ");";
-                }
-                cmd.CommandText = query; 
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-            return Annotation.IdAnnotation; 
-        }
-
-        internal StudentAnnotation GetAnnotation(int? IdAnnotation)
-        {
-            StudentAnnotation a;
-            if (IdAnnotation == null)
-                return null; 
-            a = new StudentAnnotation();
-            using (DbConnection conn = dl.Connect())
-            {
-                DbDataReader dRead;
-                DbCommand cmd = conn.CreateCommand();
-                string query = "SELECT *" +
-                    " FROM StudentsAnnotations" +
-                    " WHERE IdAnnotation=" + IdAnnotation;
-                query += ";";
-                cmd.CommandText = query;
-                dRead = cmd.ExecuteReader();
-                dRead.Read(); 
-                a = dl.GetAnnotationFromRow(dRead);
-                cmd.Dispose(); 
-            }
-            return a;
-        }
-
+       
         
         #endregion
     }
